new(modern_bpf): move `flags` utils in a dedicated folder
