+++ test/modern_bpf/flags/flags_definitions.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <sys/syscall.h>
+
+/* capabilities */
+#if defined(__NR_capget) && defined(__NR_capset)
+
+#include <sys/capability.h>
+uint64_t capabilities_to_scap(unsigned long caps);
+
+#endif /* __NR_capget */
